move definition of a couple global constants from a .h to a .cpp
